requirements.txt
kfp==2.13.0
google-cloud-aiplatform>=1.38.1,<2.0.0
google-cloud-bigquery>=3.10.0,<4.0.0
db-dtypes
protobuf>=4.21.1,<5.0.0 
google-auth>=2.14.1,<3.0.0
PyYAML>=6.0,<7.0
Jinja2>=3.1,<4.0
click>=8.0.0,<9.0.0
google-cloud-artifact-registry>=1.0.0,<2.0.0
python-dateutil